backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
supabase==2.0.3
pydantic==2.5.0
pydantic-settings==2.1.0
# AI/ML
torch==2.1.0
torchvision==0.16.0
pillow==10.1.0
numpy==1.24.3
opencv-python==4.8.1.78
# OCR & Text Analysis
pytesseract==0.3.10
textstat==0.7.3
# Image Processing
scikit-image==0.22.0
# Utilities
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
